// app/common/interfaces.ts
export interface IDialogOptions {
  /**
   * Title of the dialog window.
   */
  title: string;
  /**
   * Path where open the dialog.
   */
  defaultPath?: string;
  /**
   * Message to show in the dialog window.
   */
  message?: string;
  /**
   * Allow multiple paths to be selected.
   */
  multiple?: boolean;
  /**
   * Show the dialog as modal window.
   */
  modal?: boolean;
  /**
   * Array of file types that can be displayed or selected.
   */
  filters?: FileFilter[];
}

/**
 * Wrapper interface used to type an `IPCLogger` object
 * in the renderer process via `ContextBridge`.
 */
export interface IRendererLogger {
  info: (m: string) => void;
  warn: (m: string) => void;
  error: (m: string) => void;
}

/**
 * Wrapper interface used to type an `RendererIPCHandler`
 * object in the renderer process via `ContextBridge`.
 */
export interface IRendererIPCHandler {
  configure: () => Promise<void>;
  send: (channel: string, ...args: any[]) => void;
  receive: (channel: string, f: Function) => void;
  invoke: (channel: string, ...args: any[]) => void;
}